Various fixes
Moved all of the escaping single quotations over to DatabaseManager to
centralise it all.
Fixed issue with prices not coming up as double decimal when copying
previous entries in New Stock.
Fixed issue with prices not being double decimal in CustomersOrders.
Documentation updated.
Added feedback if there is no files to update the database from.